Added cblas

+++ F2CLIBS/libf2c/f2c.h
@@ -7,7 +7,7 @@
 #ifndef F2C_INCLUDE
 #define F2C_INCLUDE
 
-typedef long int integer;
+typedef int integer;
 typedef unsigned long int uinteger;
 typedef char *address;
 typedef short int shortint;
@@ -15,7 +15,7 @@ typedef float real;
 typedef double doublereal;
 typedef struct { real r, i; } complex;
 typedef struct { doublereal r, i; } doublecomplex;
-typedef long int logical;
+typedef int logical;
 typedef short int shortlogical;
 typedef char logical1;
 typedef char integer1;
@@ -43,8 +43,8 @@ typedef short ftnlen;
 typedef short ftnint;
 #else
 typedef long int flag;
-typedef long int ftnlen;
-typedef long int ftnint;
+typedef int ftnlen;
+typedef int ftnint;
 #endif
 
 /*external read, write*/
